app/src/main/java/com/kade/dicoding/footballapps/presenter/TeamPresenter.kt
package com.kade.dicoding.footballapps.presenter

import com.kade.dicoding.footballapps.api.ApiRepository
import com.kade.dicoding.footballapps.api.TheSportDBApi
import com.kade.dicoding.footballapps.coroutine.CoroutineContextProvider
import com.kade.dicoding.footballapps.model.TeamResponse
import com.kade.dicoding.footballapps.view.TeamView
import com.google.gson.Gson
import kotlinx.coroutines.*

class TeamPresenter(
    private val view: TeamView,
    private val apiRepository: ApiRepository,
    private val gson: Gson,
    private val context: CoroutineContextProvider = CoroutineContextProvider()
) {
    fun getTeamList(league: String?) {
        view.showLoading()

        GlobalScope.launch(context.main) {
            val data = withContext(Dispatchers.IO) {
                gson.fromJson(
                    apiRepository.doRequest(TheSportDBApi.getTeamList(league)),
                    TeamResponse::class.java
                )
            }
            view.showTeamList(data.teams)
            view.hideLoading()
        }
    }

}